+++ ibis/backends/druid/datatypes.py
@@ -7,7 +7,7 @@ from sqlalchemy.ext.compiler import compiles
 
 import ibis.expr.datatypes as dt
 from ibis.backends.base.sql.alchemy.datatypes import AlchemyType
-from ibis.backends.base.sql.glot.datatypes import DruidType as SqlglotDruidType
+from ibis.backends.base.sqlglot.datatypes import DruidType as SqlglotDruidType
 
 
 class DruidDateTime(sat.TypeDecorator):
